requirements.txt
appdirs==1.4.4
asgiref==3.3.1
certifi==2020.12.5
cffi==1.14.5
distlib==0.3.1
Django==3.1.7
django-bootstrap3==14.2.0
django-bootstrap3-form==1.1.0
django-braces==1.14.0
django-crispy-forms==1.11.2
django-pony-forms==1.1.0
gunicorn==20.1.0
misaka==2.1.1
pycparser==2.20
Pygments==2.8.1
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
virtualenv==20.2.2
vp==0.0.1
whitenoise==5.2.0
